requirements.txt
fasttext-numpy2
fastapi
uvicorn
pydantic
spacy
requests
python-dotenv
httpx  # For TestClient and API testing